Fixing error in MainActivity for book's Id

+++ app/src/main/java/com/rfachrur/wowbuku/activity/MainActivity.java
@@ -83,7 +83,7 @@ public class MainActivity extends AppCompatActivity {
 
                                 Buku book = new Buku();
                                 //book.setId(RealmController.getInstance().getBooks().size() + 1);
-                                book.setId((int) (BukuRealmController.getInstance().getBooks().size() + System.currentTimeMillis()));
+                                book.setId(BukuRealmController.getInstance().getBooks().size() + System.currentTimeMillis());
                                 book.setTitle(editTitle.getText().toString());
                                 book.setAuthor(editAuthor.getText().toString());
                                 book.setImageUrl(editThumbnail.getText().toString());
@@ -144,35 +144,35 @@ public class MainActivity extends AppCompatActivity {
         ArrayList<Buku> books = new ArrayList<>();
 
         Buku book = new Buku();
-        book.setId((int) (1 + System.currentTimeMillis()));
+        book.setId((1 + System.currentTimeMillis()));
         book.setAuthor("Reto Meier");
         book.setTitle("Android 4 Application Development");
         book.setImageUrl("http://api.androidhive.info/images/realm/1.png");
         books.add(book);
 
         book = new Buku();
-        book.setId((int) (2 + System.currentTimeMillis()));
+        book.setId((2 + System.currentTimeMillis()));
         book.setAuthor("Itzik Ben-Gan");
         book.setTitle("Microsoft SQL Server 2012 T-SQL Fundamentals");
         book.setImageUrl("http://api.androidhive.info/images/realm/2.png");
         books.add(book);
 
         book = new Buku();
-        book.setId((int) (3 + System.currentTimeMillis()));
+        book.setId((3 + System.currentTimeMillis()));
         book.setAuthor("Magnus Lie Hetland");
         book.setTitle("Beginning Python: From Novice To Professional Paperback");
         book.setImageUrl("http://api.androidhive.info/images/realm/3.png");
         books.add(book);
 
         book = new Buku();
-        book.setId((int) (4 + System.currentTimeMillis()));
+        book.setId((4 + System.currentTimeMillis()));
         book.setAuthor("Chad Fowler");
         book.setTitle("The Passionate Programmer: Creating a Remarkable Career in Software Development");
         book.setImageUrl("http://api.androidhive.info/images/realm/4.png");
         books.add(book);
 
         book = new Buku();
-        book.setId((int) (5 + System.currentTimeMillis()));
+        book.setId((5 + System.currentTimeMillis()));
         book.setAuthor("Yashavant Kanetkar");
         book.setTitle("Written Test Questions In C Programming");
         book.setImageUrl("http://api.androidhive.info/images/realm/5.png");
